Address PR #150 review: testingMode bypass, shared ValidationMixin, non-vacuous test
- widget_create empty-baseline guard now honors self.testingMode (was a plain
  confirmOkCancelDisplay that would block CI).
- Move _segmentationIsEmpty / _colorTableNotTerminology into a dedicated
  ValidationMixin (MorphoDepotWidget inherits it) instead of CreateTabMixin reaching
  into ReleaseTabMixin -- the shared dependency is now explicit.
- stress_continuous_colortable_guard asserts GenericAnatomyColors is present so the
  positive case can't pass vacuously.

Net 15/15.
