+++ MorphoDepot/MorphoDepotLib/widget_create.py
@@ -539,9 +539,9 @@ class CreateTabMixin:
         # Part E bad-input guard: a selected-but-empty baseline segmentation would stage an empty
         # baseline. (Color-table terminology is validated below for archival / warned for short-term.)
         if self._segmentationIsEmpty(sourceSegmentation):
-            if not slicer.util.confirmOkCancelDisplay(
+            if not (self.testingMode or slicer.util.confirmOkCancelDisplay(
                     "The selected baseline segmentation has no segments. Stage it anyway (the "
-                    "repository would start with an empty baseline)?", windowTitle="Empty baseline"):
+                    "repository would start with an empty baseline)?", windowTitle="Empty baseline")):
                 return None
 
         accessionData = self.createUI.accessionForm.accessionData()
